rename select component
